Bugfix: Relaunch Container
The relaunch container relied on a local change in the orignal OC objects. The bug is fixed by not including the original container and implementing the code in this object again. This may not be a good practice, but it makes the object less prone to accidental changes.

(cherry picked from commit 058c2297aeaf7b2e6b810c55677ac9161ba1a4be)